requirements.txt
telethon
pymongo
python-dotenv
streamlit
pandas
